exract interface from GlobalAnalayzer
