+++ core/src/main/java/edu/ucr/cs/riple/core/global/GlobalAnalyzer.java
@@ -24,31 +24,12 @@
 
 package edu.ucr.cs.riple.core.global;
 
-import com.google.common.collect.ImmutableMultimap;
 import com.google.common.collect.ImmutableSet;
-import com.google.common.collect.Multimaps;
-import edu.ucr.cs.riple.core.Config;
-import edu.ucr.cs.riple.core.ModuleInfo;
-import edu.ucr.cs.riple.core.Report;
-import edu.ucr.cs.riple.core.injectors.VirtualInjector;
-import edu.ucr.cs.riple.core.metadata.field.FieldDeclarationAnalysis;
-import edu.ucr.cs.riple.core.metadata.index.Bank;
 import edu.ucr.cs.riple.core.metadata.index.Error;
 import edu.ucr.cs.riple.core.metadata.index.Fix;
-import edu.ucr.cs.riple.core.metadata.method.MethodDeclarationTree;
-import edu.ucr.cs.riple.core.metadata.method.MethodNode;
-import edu.ucr.cs.riple.core.metadata.trackers.MethodRegionTracker;
-import edu.ucr.cs.riple.core.util.Utility;
-import edu.ucr.cs.riple.injector.changes.AddAnnotation;
-import edu.ucr.cs.riple.injector.location.OnMethod;
 import edu.ucr.cs.riple.injector.location.OnParameter;
-import java.util.Collections;
 import java.util.List;
-import java.util.Optional;
-import java.util.OptionalInt;
 import java.util.Set;
-import java.util.stream.Stream;
-import javax.annotation.Nullable;
 
 /**
  * Analyzer for downstream dependencies.
@@ -59,128 +40,10 @@ import javax.annotation.Nullable;
  * of additional errors) of each such change, by summing across all downstream dependencies. This
  * data then be fed to the Annotator main process in the decision process.
  */
-public class GlobalAnalyzer {
-
-  /** Set of downstream dependencies. */
-  private final ImmutableSet<ModuleInfo> modules;
-  /** Public APIs in the target modules that have a non-primitive return value. */
-  private final ImmutableMultimap<Integer, MethodImpact> methods;
-  /** Annotator Config. */
-  private final Config config;
-  /** Method declaration tree instance. */
-  private final MethodDeclarationTree tree;
-  /**
-   * VirtualInjector instance. Annotations should only be loaded in a library model and not
-   * physically injected.
-   */
-  private final VirtualInjector injector;
-
-  public GlobalAnalyzer(Config config, MethodDeclarationTree tree) {
-    this.config = config;
-    this.modules = config.downstreamInfo;
-    this.tree = tree;
-    this.injector = new VirtualInjector(config);
-    this.methods =
-        Multimaps.index(
-            tree.getPublicMethodsWithNonPrimitivesReturn().stream()
-                .map(MethodImpact::new)
-                .collect(ImmutableSet.toImmutableSet()),
-            MethodImpact::hashCode);
-  }
+public interface GlobalAnalyzer {
 
   /** Analyzes effects of changes in public methods in downstream dependencies. */
-  public void analyzeDownstreamDependencies() {
-    System.out.println("Analysing downstream dependencies...");
-    Utility.setScannerCheckerActivation(modules, true);
-    Utility.buildDownstreamDependencies(config);
-    Utility.setScannerCheckerActivation(modules, false);
-    // Collect callers of public APIs in module.
-    MethodRegionTracker tracker = new MethodRegionTracker(config.downstreamInfo, tree);
-    // Generate fixes corresponding methods.
-    ImmutableSet<Fix> fixes =
-        methods.values().stream()
-            .filter(
-                input ->
-                    !tracker
-                        .getCallersOfMethod(input.node.location.clazz, input.node.location.method)
-                        .isEmpty()) // skip methods that are not called anywhere.
-            .map(
-                methodImpact ->
-                    new Fix(
-                        new AddAnnotation(
-                            new OnMethod(
-                                "null",
-                                methodImpact.node.location.clazz,
-                                methodImpact.node.location.method),
-                            config.nullableAnnot),
-                        "null",
-                        "null",
-                        "null",
-                        true))
-            .collect(ImmutableSet.toImmutableSet());
-    // Explorer initializations.
-    FieldDeclarationAnalysis fieldDeclarationAnalysis =
-        new FieldDeclarationAnalysis(config.downstreamInfo);
-    Bank<Error> errorBank =
-        new Bank<>(
-            config.downstreamInfo.stream()
-                .map(info -> info.dir.resolve("errors.tsv"))
-                .collect(ImmutableSet.toImmutableSet()),
-            Error::new);
-    Bank<Fix> fixBank =
-        new Bank<>(
-            config.downstreamInfo.stream()
-                .map(info -> info.dir.resolve("fixes.tsv"))
-                .collect(ImmutableSet.toImmutableSet()),
-            Fix.factory(config, fieldDeclarationAnalysis));
-    DownstreamImpactAnalyzer analyzer =
-        new DownstreamImpactAnalyzer(injector, errorBank, fixBank, tracker, fixes, tree, 1, config);
-    ImmutableSet<Report> reports = analyzer.explore();
-    // Update method status based on the results.
-    methods
-        .values()
-        .forEach(
-            method -> {
-              MethodNode node = method.node;
-              Set<OnParameter> impactedParameters = analyzer.getImpactedParameters(node.location);
-              reports.stream()
-                  .filter(input -> input.root.toMethod().equals(node.location))
-                  .findAny()
-                  .ifPresent(report -> method.setStatus(report, impactedParameters));
-            });
-    System.out.println("Analysing downstream dependencies completed!");
-  }
-
-  /**
-   * Retrieves the corresponding {@link MethodImpact} to a fix.
-   *
-   * @param fix Target fix.
-   * @return Corresponding {@link MethodImpact}, null if not located.
-   */
-  @Nullable
-  private MethodImpact fetchStatus(Fix fix) {
-    if (!fix.isOnMethod()) {
-      return null;
-    }
-    OnMethod onMethod = fix.toMethod();
-    int predictedHash = MethodImpact.hash(onMethod.method, onMethod.clazz);
-    Optional<MethodImpact> optional =
-        this.methods.get(predictedHash).stream()
-            .filter(m -> m.node.location.equals(onMethod))
-            .findAny();
-    return optional.orElse(null);
-  }
-
-  /**
-   * Returns the effect of applying a fix on the target on downstream dependencies.
-   *
-   * @param fix Fix targeting an element in target.
-   * @return Effect on downstream dependencies.
-   */
-  private int effectOnDownstreamDependencies(Fix fix) {
-    MethodImpact status = fetchStatus(fix);
-    return status == null ? 0 : status.getEffect();
-  }
+  void analyzeDownstreamDependencies();
 
   /**
    * Returns the lower bound of number of errors of applying a fix and its associated chain of fixes
@@ -189,14 +52,7 @@ public class GlobalAnalyzer {
    * @param tree Tree of the fix tree associated to root.
    * @return Lower bound of number of errors on downstream dependencies.
    */
-  public int computeLowerBoundOfNumberOfErrors(Set<Fix> tree) {
-    OptionalInt lowerBoundEffectOfChainOptional =
-        tree.stream().mapToInt(this::effectOnDownstreamDependencies).max();
-    if (lowerBoundEffectOfChainOptional.isEmpty()) {
-      return 0;
-    }
-    return lowerBoundEffectOfChainOptional.getAsInt();
-  }
+  int computeLowerBoundOfNumberOfErrors(Set<Fix> tree);
 
   /**
    * Returns the upper bound of number of errors of applying a fix and its associated chain of fixes
@@ -205,9 +61,7 @@ public class GlobalAnalyzer {
    * @param tree Tree of the fix tree associated to root.
    * @return Lower bound of number of errors on downstream dependencies.
    */
-  public int computeUpperBoundOfNumberOfErrors(Set<Fix> tree) {
-    return tree.stream().mapToInt(this::effectOnDownstreamDependencies).sum();
-  }
+  int computeUpperBoundOfNumberOfErrors(Set<Fix> tree);
 
   /**
    * Returns set of parameters that will receive {@code @Nullable}, if any of the methods in the
@@ -216,16 +70,7 @@ public class GlobalAnalyzer {
    * @param fixTree Fix tree.
    * @return Immutable set of impacted parameters.
    */
-  public ImmutableSet<OnParameter> getImpactedParameters(Set<Fix> fixTree) {
-    return fixTree.stream()
-        .filter(Fix::isOnMethod)
-        .flatMap(
-            fix -> {
-              MethodImpact status = fetchStatus(fix);
-              return status == null ? Stream.of() : status.getImpactedParameters().stream();
-            })
-        .collect(ImmutableSet.toImmutableSet());
-  }
+  ImmutableSet<OnParameter> getImpactedParameters(Set<Fix> fixTree);
 
   /**
    * Collects list of triggered errors in downstream dependencies if fix is applied in the target
@@ -234,24 +79,12 @@ public class GlobalAnalyzer {
    * @param fix Fix instance to be applied in the target module.
    * @return List of triggered errors.
    */
-  public List<Error> getTriggeredErrors(Fix fix) {
-    // We currently only store impact of methods on downstream dependencies.
-    if (!fix.isOnMethod()) {
-      return Collections.emptyList();
-    }
-    MethodImpact impact = fetchStatus(fix);
-    if (impact == null) {
-      return Collections.emptyList();
-    }
-    return impact.getTriggeredErrors();
-  }
+  List<Error> getTriggeredErrors(Fix fix);
 
   /**
    * Updates state of methods after injection of fixes in target module.
    *
    * @param fixes Set of injected fixes.
    */
-  public void updateImpactsAfterInjection(Set<Fix> fixes) {
-    this.methods.values().forEach(methodImpact -> methodImpact.updateStatus(fixes));
-  }
+  void updateImpactsAfterInjection(Set<Fix> fixes);
 }
